+++ changelog.md
@@ -5,6 +5,7 @@ Features
 ---------
 * Add a `--batch` option as an alternative to STDIN.
 * Deprecate `$MYSQL_UNIX_PORT` environment variable in favor of `$MYSQL_UNIX_SOCKET`.
+* Support `--username` and `$MYSQL_USER` to set username.
 
 
 Bug Fixes